Feat: RTR
Task: S10P12D203-52

+++ Backend/tranvel/src/main/java/com/ssafy/tranvel/dto/InquiryDto.java
@@ -1,7 +1,5 @@
 package com.ssafy.tranvel.dto;
 
-
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
@@ -9,9 +7,9 @@ import lombok.Setter;
 @Getter @Setter
 public class InquiryDto {
     @NotNull
-    private int userId;
+    private Long userId;
 
-    private int inquiryId;
+    private Long inquiryId;
 
     private String title;
